Changed class name

+++ tests/test_util.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include "./test_util.hpp"
 
-void checkEqual(int actual, int expected, std::string file, int line) {
+void checkEqual(long int actual, long int expected, std::string file, int line) {
   if (actual != expected) {
     std::cerr << file << ":" << line << ": ";
     std::cerr << "Expected " << actual << " to equal " << expected << "." << std::endl;
