+++ tests/test_util.hpp
@@ -5,6 +5,6 @@
 
 #define CHECK_EQUAL(actual, expected) checkEqual((actual), (expected),  __FILE__, __LINE__);
 
-void checkEqual(int actual, int expected, std::string file, int line);
+void checkEqual(long int actual, long int expected, std::string file, int line);
 
 #endif // TESTS_UTIL_H
